Allocation issue
Discovered an issue when allocating the pages for the physical memory
map. The calculation for how many pages would need to be allocated to
accomdate the ARC_MMap structure that is passed to the kernel rounded
down, allocating a total of 0 contiguous pages. This returned NULL by
the allocator and resulted in a write to 0x0. This has been fixed with a
simple round up.