ARC_MMap construction modifications
There is now logic implemented to modify the base GRUB memory map in
order to mark all memory map entries before the end of the bootstrapper
as reserved. This also includes logic for splitting a memory map entry
in order to maintain as much free memory as possible while keeping the
bootstrapper safe.