Mega rework
Reworked basically the entire bootstrapper. The code is now much cleaner
and straight forward to read. Comments in some sections may be a little
bit out dated, but I will update those in another commit. However, this
is a working bootstrapper as it does get to the kernel. There has been a
rework with the freelist allocator implementation which the kernel will
need to be updated with as well.