Some fixes
Added some more comments to mbparse.c and cleaned them up in main.c.
There was quite a severe oversight in the freelist implementation where
32-bit void pointers would be cast to 64-bit unsigned integers. This
resulted in the upper 32-bits of addresses being set to 0xFFFF.FFFF
which caused problems when calculating the size of the region. Currently
there is an issue on hardware where the freelist gets corrupted in some
fashion at 0x6E6000 resulting in a failure to allocate the HHDM, I still
don't know what the root of this problem is. Strangely this only happens
if the bootstrapper is loaded at 1MB, as if it is loaded at 2MB all C
code seems to execute fine, though the kernel is not reached or a crash
is experienced in the kernel.