PMM, Framebuffer: changed some functionality
The first change comes in the PMM. All addresses within the freelists
initialized by the PMM are now in the HHDM. This saves some time for the
kernel and is possible because of the freelist implementation using
uint64_t's. Since adding HHDM_VADDR to any physical address only effects
the upper 32-bits, the bootstrapper can still allocate page frames
without problem. The framebuffer is now also mapped in for the kernel so
that as soon as the kernel starts, it can start blitting out data for
the user to observe immediately (if it wants to). Lastly debug messages
have been changed so that errors are reported no matter of the presence
of debug.enable, and have received colors.