Freelist: rounding error
The freelist implementation has been working, despite the fact that it
should not have. There was a severe problem that has only showed itself
recently in the kernel. Basically, the metadata and first allocation
would collide, leading to a scenario where an file descriptor's node
pointer would be corrupted by one as it just so happened to land exactly
on meta->free_objects. This would then cascade down leading to a general
protection fault.