Fix: terminal, CPUID,  pager
There was an issue with print statements done prior to terminal initialization not showing up in the serial output. This is because a check is done to ensure that there has been a terminal set prior to printing to the serial console. To resolve this the print to serial has been moved above this check. The CPUID function was not checking important paging features, those checks have now been added. A config.h file has been added to a newly created arch/x86-64 include folder; this is due to some reworks happening inside of the kernel. Finally the x86-64 pager, which is kept under x86 (should really be moved to be implemented under x86-64), has had an ordering issue resolved. It was placing GiB page entries as PML4 entries, and 2 MiB page entries as PML3 entries. This was not right and they needed to be shifted down by one table entry. Things are now done properly.